Fix: Refactor the llm.py file in the project root
